+++ cargo-workspaces/src/utils/mod.rs
@@ -6,7 +6,7 @@ mod listable;
 mod pkg;
 mod version;
 
-pub use cargo::{cargo, change_versions};
+pub use cargo::{cargo, change_versions, check_index};
 pub use changable::{ChangeData, ChangeOpt};
 pub(crate) use error::{debug, info};
 pub use error::{get_debug, set_debug, Error, GREEN, MAGENTA, TERM_ERR, TERM_OUT};
@@ -17,4 +17,4 @@ pub use version::VersionOpt;
 
 pub type Result<T = ()> = std::result::Result<T, Error>;
 
-pub const INTERNAL_ERR: &'static str = "Internal error message. Please create an issue on https://github.com/pksunkara/cargo-workspaces";
+pub const INTERNAL_ERR: &str = "Internal error message. Please create an issue on https://github.com/pksunkara/cargo-workspaces";